% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_numeric_category.R
\name{boxplot_numeric_category}
\alias{boxplot_numeric_category}
\title{Multiple Boxplot for a Numeric Variable across Categories}
\usage{
boxplot_numeric_category(dataframe, x, y)
}
\arguments{
\item{dataframe}{The dataframe that contains the variables that you would like to use to create the boxplot. The class of this parameter must be dataframe.}

\item{x}{The categorical variable containing the categories that we want to plot the distribution of a numeric variable across. x must be a vector of class character or factor.}

\item{y}{The numeric variable that we want to examine the distribution of across a categorical variable. y must be a vector of class double, integer, or numeric.}
}
\value{
A single panel multiple boxplot showing the distribution of the variable y in each category contained in the variable x.
}
\description{
This function creates a multiple boxplot that shows the distribution of a numeric variable in each different category contained in a categorical variable. This function will remove any NA values in the data prior to creation of the boxplot.
}
\examples{
## Using the "CO2" dataset from the r "datasets" package,
## create a boxplot that shows the distribution of CO2 concentration
## of grass plants in each treatment type (chilled or nonchilled)
boxplot_numeric_category(dataframe = datasets::CO2, x = Treatment, y = conc)

## Using the "starwars" dataset from the dplyr package,
## create a boxplot that shows the distribution of the height
## of starwars characters based on the sex of the characters
dplyr::starwars \%>\% boxplot_numeric_category(sex, height)
}
